BIOS ROM changed to Block RAM Generator IP.
The BIOS ROM was changed to a Xilinx Block RAM Generator IP to
fascilitate changing the ROM BIOS in the .bit file without having to
re-synthesize and re-implement the entire design if the only thing that
changed is the ROM code.

Two utilities were created to generate .coe and .mem files from the
assembled ROM BIOS: bin2coe.sh and bin2mem.sh. The makefile was updated
to incorporate the utilites in the build process. The .coe file will
only be utilized when the Block RAM Generator IP is re-synthesized. The
.mem file is for the updatemem utility to update the .bit file.

The file bios.mmi is also used by updatemem in order to find the block
RAM in the .bit file. Everytime the design is re-implemented, you must
verify that the location of the block RAMs have not changed.

updatebit.bat was added to simplify updating the .bit file from the
Windows command prompt.

The .gitignore file was updated to ignore the .mem and .coe files, as
well as files generated from updatemem.