Added itoa routine, rewrote memory map, and reorganized.
Created an itoa routine to convert a 16-bit integer into a NUL
terminated string.

I had to rewrite the memory map to include a BSS segment to create the
integer string. Additionally, the memory.conf file now only includes the
memory locations that the BIOS cares about. This includes the ROM area,
and a BIOS RAM area ($02FF-$0300).

Deleted the include files that would import symbols. Created a new
conio.inc that uses macros for cursor_on and cursor_off. The reset
interrupt now goes to the start: label, which tests for RAM before
proceding to the main: label.