Partial PS/2 Keyboard module implemented.
Implemented a receive-only PS/2 keyboard interface. The user interacts
by polling the STATUS register to check if a scan code is available.
Reading the SCANCODE register will reset the DATA_RDY flag in the
STATUS register. The BIOS was updated to comment out the serial monitor
and continuously poll the keyboard, outputing the scan code to the
seven-segment display.