ROM monitor now uses PS/2 keyboard instead of UART.
The ROM monitor interface was rewritten to use the keyboard. Like the
UART monitor, an escape key will clear the line, a tab key is ignored,
and backspace will delete previously entered characters.

The scan code keymap was updated to include shifted ASCII characters.

Additionally, the VGA initialization function was removed and added
to the main ROM program. The initialization was mostly rewritten into
the clrscr subroutine.

All functions are working implementations, but not optimized for speed
or space.